Big refactor ready for packing, hopefully?

+++ index_data_files.py
@@ -2,20 +2,27 @@ import argparse
 import json
 import os
 import struct
+import re
 import subprocess
 from collections import defaultdict
 from multiprocessing import Pool, cpu_count
 
 
+def byte_pad_string(string):
+    string_bytes = string.encode('ascii') + b'\x00'
+    padding_length = (4 - (len(string_bytes) % 4)) % 4
+    return string_bytes + b'\x00' * padding_length
+
+
 class FileEntry:
-    def __init__(self, file_entry=None, filename=None):
-        self.file_entry = file_entry or (0, 0, 0, 0, 0)  # default 5-tuple
-        self.filename = filename or ""
+    def __init__(self, filename, file_entry=None):
+        self.filename = filename
+        self.file_entry = file_entry or [0] * 6
         self.real_uncompressed_size = None
         self.real_compressed_size = None
     
     def entry_size(self):
-        return self.file_entry[0]
+        return 16 + len(byte_pad_string(self.filename))
     
     def offset(self):
         return self.file_entry[1]
@@ -23,17 +30,23 @@ class FileEntry:
     def update_from_data_file(self, data_file):
         # For some reason, certain files are marked as size 0 in the index, even though they exist
         # in the data file. In these cases, we need to check the data file itself.
-        if self.file_entry[3] != 0 and self.file_entry[2] != 0:
+        if not self.is_ghost_file():
             return
 
         data_file.seek(self.offset())
-        if data_file.read(4) == b"\x52\x4E\x42\x01":  # "RNC 1"
-            self.real_uncompressed_size = self.struct.unpack_from('>I', data_file.read(4))[0]
-            self.real_compressed_size = self.struct.unpack_from('>I', data_file.read(4))[0]
+        header = data_file.read(4)
+        if header == b"RNC\x01":  # RNC compression
+            self.real_uncompressed_size = struct.unpack_from('>I', data_file.read(4))[0]
+            print(f"Real uncompressed size of {self.filename}: {self.real_uncompressed_size}")
+            self.real_compressed_size = struct.unpack_from('>I', data_file.read(4))[0]
+            print(f"Real compressed size of {self.filename}: {self.real_compressed_size}")
         else:
             # Fuck knows
             self.real_compressed_size = self.real_uncompressed_size = 0x800
 
+    def is_ghost_file(self):
+        return self.file_entry[2] == 0 or self.file_entry[3] == 0
+    
     def _file_size(self, compressed):
         size = self.file_entry[3 if compressed else 2]
         if size == 0:
@@ -57,10 +70,7 @@ class FileEntry:
     def write_to_file(self):
         # Serialize the file entry and filename
         packed_entry = struct.pack('<IIIHBB', *self.file_entry)
-        filename_bytes = self.filename.encode('ascii') + b'\x00'
-        padding_length = (4 - (len(filename_bytes) % 4)) % 4
-        filename_bytes += b'\x00' * padding_length
-        return packed_entry + filename_bytes
+        return packed_entry + byte_pad_string(self.filename)
 
     @classmethod
     def read_from_file(cls, data, index):
@@ -75,34 +85,79 @@ class FileEntry:
         filename = entry_data[16:].decode('ascii').rstrip('\x00')
 
         return cls(file_entry=file_entry, filename=filename), index
+    
+    @staticmethod
+    def compress_file(path):
+        temp_dat_path = path + '.dat'
+        command = [
+            'rnc_lib.exe', 'p', path, temp_dat_path, '-m=1'
+        ]
+        subprocess.check_call(command)#, stdout=subprocess.DEVNULL)
+        return temp_dat_path
+    
+    @classmethod
+    def compressed_size_from_path(cls, path, entry_metadata):
+        if entry_metadata.get("is_compressed", True):
+            temp_compressed_path = cls.compress_file(path)
+            size = os.path.getsize(temp_compressed_path)
+            os.remove(temp_compressed_path)
+            return size
+        else:
+            return os.path.getsize(path)
+
+    @classmethod
+    def from_path(cls, path, entry_metadata):
+        assert os.path.isfile(path)
+        filename = os.path.basename(path)
+
+        # TODO: this "file_entry" stuff needs refactoring
+        file_entry = [
+            16 + len(byte_pad_string(filename)),  # Entry size
+            entry_metadata["offset"],  # TODO: calculate the offset as needed when writing the data file
+            os.path.getsize(path),  # Uncompressed size
+            cls.compressed_size_from_path(path, entry_metadata),  # Compressed size
+            entry_metadata.get("flags", 0),  # Flags(?)
+            0x80 if entry_metadata.get("is_compressed", True) else 0x00  # Is compressed
+        ]
+
+        entry = FileEntry(filename, file_entry)
+
+        if entry_metadata.get("is_ghost_file", False):
+            entry.real_uncompressed_size = file_entry[2]
+            entry.real_compressed_size = file_entry[3]
+            entry.file_entry[2] = entry.file_entry[3] = 0
+        return entry
 
 
 class IndexSection:
     UNKNOWN_SECTION_NAME = "_UNKNOWN"
 
-    def __init__(self, section_size=0, unknown=0, section_name=None, entries=None):
-        self.section_size = section_size
+    def __init__(self, unknown=0, section_name=None, entries=None):
         self.unknown = unknown
         self._section_name = section_name
         self.entries = entries or []
     
     def extracted_dir_name(self):
         return self._section_name or self.UNKNOWN_SECTION_NAME
+    
+    def section_name_padded(self):
+        if self._section_name:
+            return byte_pad_string(self._section_name)
+        else:
+            return b""
+    
+    def section_size(self):
+        return 12 + len(self.section_name_padded())
 
     def write_to_file(self):
-        if self.section_size == 12:  # No name
-            packed_header = struct.pack('<III', self.section_size, len(self.entries), self.unknown)
-            section_name_bytes = b""
+        if not self._section_name:  # NOTE: number of entries and unknown are swapped for some reason
+            packed_header = struct.pack('<III', self.section_size(), len(self.entries), self.unknown)
         else:
-            assert self._section_name
-            packed_header = struct.pack('<III', self.section_size, self.unknown, len(self.entries))
-            section_name_bytes = self._section_name.encode('ascii') + b'\x00'
-            padding_length = (4 - (len(section_name_bytes) % 4)) % 4
-            section_name_bytes += b'\x00' * padding_length
+            packed_header = struct.pack('<III', self.section_size(), self.unknown, len(self.entries))
 
         # Serialize each file entry
         entry_data = b''.join(entry.write_to_file() for entry in self.entries)
-        return packed_header + section_name_bytes + entry_data
+        return packed_header + self.section_name_padded() + entry_data
 
     @classmethod
     def read_from_file(cls, data, index):
@@ -127,7 +182,28 @@ class IndexSection:
             entry, index = FileEntry.read_from_file(data, index)
             entries.append(entry)
 
-        return cls(section_size=section_size, unknown=unknown, section_name=section_name, entries=entries), index
+        return cls(unknown=unknown, section_name=section_name, entries=entries), index
+    
+    @classmethod
+    def from_path(cls, path, section_metadata):
+        # Replicate the filename ordering of the original index file
+        def filename_key(filename):
+            filename = filename.lower().replace("_", "z")
+            parts = re.findall(r'\d+|\D', filename)
+            def number_if_possible(s):
+                try:
+                    return str(int(s)).zfill(10)
+                except ValueError:
+                    return s
+            return tuple(number_if_possible(p) for p in parts)
+
+        entries = []
+        for filename in sorted(os.listdir(path), key=filename_key):
+            filepath = os.path.join(path, filename)
+            if not os.path.isfile(filepath) or filename.startswith("."):
+                continue
+            entries.append(FileEntry.from_path(filepath, section_metadata["entries"].get(filename)))
+        return cls(section_metadata.get("unknown", 0), section_metadata["name"], entries)
 
 
 class AssetIndex:
@@ -159,18 +235,10 @@ class AssetIndex:
             metadata = json.load(json_file)
 
         sections = []
-        for section_data in metadata:
-            entries = [
-                FileEntry(file_entry=entry['file_entry'], filename=entry['filename'])
-                for entry in section_data['entries']
-            ]
-            section = IndexSection(
-                section_size=section_data['type'],
-                unknown=section_data['unknown'],
-                section_name=section_data['name'],
-                entries=entries
-            )
-            sections.append(section)
+        for section_metadata in metadata:
+            section_path = os.path.join(asset_dir, section_metadata["name"])
+            assert os.path.isdir(section_path)
+            sections.append(IndexSection.from_path(section_path, section_metadata))
 
         return cls(sections)
 
@@ -180,7 +248,7 @@ class AssetData:
         self.asset_index = asset_index
         self.dat_file_path = dat_file_path
 
-    def extract_files(self, output_dir):
+    def extract_files(self, output_dir, metadata_only=False):
         os.makedirs(output_dir, exist_ok=True)
 
         sections_metadata = []
@@ -202,23 +270,30 @@ class AssetData:
                     # Collect metadata needed to reconstruct the idx file
                     if not entry.is_compressed():
                         entries_metadata[entry.filename]["is_compressed"] = False
+                    if entry.is_ghost_file():
+                        entries_metadata[entry.filename]["is_ghost_file"] = True
                     if entry.flags():
                         entries_metadata[entry.filename]["flags"] = entry.flags()
+
+                    # TODO: calculate the offset as needed when writing the data file
+                    entries_metadata[entry.filename]["offset"] = entry.offset()
                 
-                sections_metadata.append({
+                section_item = {
                     'name': section.extracted_dir_name(),
-                    'size': section.section_size,  # TODO: this shouldn't be necessary
-                    'unknown': section.unknown,
                     'entries': entries_metadata
-                })
+                }
+                if section.unknown:
+                    section_item["unknown"] = section.unknown
+                sections_metadata.append(section_item)
 
         # Save metadata to JSON
         with open(os.path.join(output_dir, 'metadata.json'), 'w') as json_file:
             json.dump(sections_metadata, json_file, indent=4)
 
         # Run extraction tasks in parallel using multiprocessing
-        with Pool(cpu_count()) as pool:
-            pool.map(self._extract_file, tasks)
+        if not metadata_only:
+            with Pool(cpu_count()) as pool:
+                pool.map(self._extract_file, tasks)
 
     def _extract_file(self, args):
         dat_file_path, output_file_path, entry = args
@@ -239,25 +314,23 @@ class AssetData:
         for section in self.asset_index.sections:
             section_dir = os.path.join(asset_dir, section.extracted_dir_name())
 
-            for entry in section.entries:
-                input_file_path = os.path.join(section_dir, entry.filename)
+            with open(self.dat_file_path, 'wb') as dat_file:
+                for entry in section.entries:
+                    input_file_path = os.path.join(section_dir, entry.filename)
 
-                temp_dat_path = input_file_path + '.dat'
-                command = [
-                    'rnc_lib.exe', 'p', input_file_path, temp_dat_path, '-m=1'
-                ]
-                subprocess.check_call(command, stdout=subprocess.DEVNULL)
+                    if entry.is_compressed():
+                        temp_dat_path = FileEntry.compress_file(input_file_path)
+                        with open(temp_dat_path, 'rb') as temp_file:
+                            packed_data = temp_file.read()
+                        os.remove(temp_dat_path)
+                    else:
+                        with open(input_file_path, 'rb') as temp_file:
+                            packed_data = temp_file.read()
 
-                with open(temp_dat_path, 'rb') as temp_file:
-                    packed_data = temp_file.read()
-
-                with open(self.dat_file_path, 'r+b') as dat_file:
                     dat_file.seek(entry.offset())
                     dat_file.write(packed_data)
 
-                os.remove(temp_dat_path)
-
-                print(f'Packed {entry.filename} to {self.dat_file_path} at offset {entry.offset():X}.')
+                    print(f'Packed {entry.filename} to {self.dat_file_path} at offset {entry.offset():X}.')
 
     @classmethod
     def from_asset_dir(cls, asset_dir, dat_file_path):
@@ -289,8 +362,8 @@ if __name__ == "__main__":
         asset_data.extract_files(args.output_dir)
     elif args.command == "pack":
         asset_data = AssetData.from_asset_dir(args.asset_dir, args.data_file)
-        asset_data.pack_files()
         with open(args.index_file, 'wb') as f:
             f.write(asset_data.asset_index.write_to_file())
+        asset_data.pack_files()
     else:
         raise argparse.ArgumentError("Unknown command!")
